+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/di/KoinInitializer.kt
@@ -2,7 +2,7 @@ package com.poli.health.aquamate.di
 
 import com.poli.health.aquamate.framework.platformModule
 import com.poli.health.aquamate.onboarding.auth.di.authModule
-import com.poli.health.aquamate.onboarding.auth.di.intakeModule
+import com.poli.health.aquamate.intake.di.intakeModule
 import com.poli.health.aquamate.onboarding.profile.di.profileModule
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration