+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 ## Version: 1.0.0
 
 ### Added
-
 - The theme that this application will use has been implemented.
 - Implemented Koin to dependency injection and Firebase with auth library.
 - Implement auth module (sign in, sign up, sign out) in the app.
@@ -14,3 +13,6 @@
 - Unify logic to inject dependencies modules in Android and iOS
 - Implement screen to user profile. This feature provides the info to calculate the daily hydration
   goal.
+
+### Fixed
+- Decouple intake feature from onboarding.